app/vlstorage: avoid binary search for the time range with the last N logs when the time range is smaller than a few seconds
The binary search over a small time range such as a few seconds requires ~30 iterations
until it converges to nanosecond precision. Such small time ranges usually contain small number
of logs even on highly loaded systems with hundreds of thousands of logs ingested per second.
So it is faster to run the query with the `| sort by (_time desc) limit N` pipe at the end
on this time range instead of running ~30 iterations for the binary search.

This shaves off more than a half of iterations for the binary search in typical production cases,
and improves the `last N logs` query performance by ~2x on average (e.g. when the limit=N query arg is passed to /select/logsql/query).

Thanks to @yuriydzobak for the comment at https://github.com/VictoriaMetrics/VictoriaLogs/issues/72#issuecomment-3317619894 ,
which helped identifying and lifting this performance bottleneck.